requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.5.0
tensorflow>=2.12.0
tensorflow-hub>=0.13.0
onnx>=1.13.0
onnx-tf>=1.10.0 